requirements.txt
asgiref==3.5.2
Django==4.0.6
django-ninja==0.19.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
idna==3.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pydantic==1.9.1
python-jose==3.3.0
pytz==2022.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
